ajout d'un fichier commun pour eviter de reecrire 4 fois les mêmes fonctions
