+++ bookingCom.py
@@ -1,22 +1,7 @@
 import time
 import csv
 from selenium import webdriver
-from geopy.geocoders import Nominatim
-
-monthCorrespondances = {
-    "01": "janvier",
-    "02": "fevrier",
-    "03": "mars",
-    "04": "avril",
-    "05": "mai",
-    "06": "juin",
-    "07": "juillet",
-    "08": "aout",
-    "09": "septembre",
-    "10": "octobre",
-    "11": "novembre",
-    "12": "decembre",
-}
+import commonFunctions as cf
 
 
 def searchCity(city):
@@ -25,15 +10,6 @@ def searchCity(city):
 
     time.sleep(2)
     driver.find_element(by="class name", value="sb-searchbox__button ").click()
-
-
-def separateDate(date):
-    """
-    :param date: dd/MM/yyyy
-    """
-    day, month, year = date.split("/")
-    month = monthCorrespondances[month]
-    return day, month, year
 
 
 def getByXpath(xpath):
@@ -86,8 +62,8 @@ def setDate(startDate, endDate):
     """
     :param arrivalDate: dd/MM/yyyy
     """
-    startDay, startMonth, startYear = separateDate(startDate)
-    endDay, endMonth, endYear = separateDate(endDate)
+    startDay, startMonth, startYear = cf.separateDate(startDate)
+    endDay, endMonth, endYear = cf.separateDate(endDate)
 
     setGoodMonthYear(startMonth, startYear)
 
@@ -99,11 +75,6 @@ def setDate(startDate, endDate):
 
     selectDay(endDay)
 
-def getLocalisationFromAdd(add):
-    geolocator = Nominatim(user_agent="main")
-    location = geolocator.geocode(add)
-    return [location.latitude, location.longitude]
-
 
 def getHotels():
     time.sleep(2)
@@ -114,9 +85,10 @@ def getHotels():
                       driver.find_elements(by="xpath", value="//div[contains(@class, '_9c5f726ff bd528f9ea6')]")))
     prices = list(map(lambda price: price.text.split(" ")[1],
                       driver.find_elements(by="xpath", value="//span[contains(@class, 'fde444d7ef _e885fdc12')]")))
-    localisations = list(map(lambda address: getLocalisationFromAdd(address.text), driver.find_elements(by="xpath", value="//span[contains(@data-testid, 'address')]")))
+    localisations = list(map(lambda address: cf.getLocalisationFromAdd(address.text),
+                             driver.find_elements(by="xpath", value="//span[contains(@data-testid, 'address')]")))
 
-    return names, grades, prices, localisations, links
+    return [names, grades, prices, localisations, links]
 
 
 def applyFamilyAndDate():
@@ -131,13 +103,6 @@ def changePage():
     driver.find_element(by="xpath", value="//button[contains(@aria-label, 'Page suivante')]").click()
 
 
-def appendToCsv(names, grades, prices, links, localisations, file):
-    file = open(file, "a")
-    for i in range(len(names) - 1):
-        csv.writer(file).writerow([names[i], grades[i], prices[i], localisations[i], links[i]])
-    file.close()
-
-
 if __name__ == '__main__':
     driver = webdriver.Firefox()
     driver.get(
@@ -148,14 +113,9 @@ if __name__ == '__main__':
     setDate("20/05/2022", "23/05/2022")
     applyFamilyAndDate()
 
-    f = open('bookingCom.csv', 'w')
-    headCsv = ["name", "grade", "price", "localisation", "link"]
-    writer = csv.writer(f)
-    writer.writerow(headCsv)
-    f.close()
+    cf.createCsv(["name", "grade", "price", "localisation", "link"], 'bookingCom.csv')
 
     while True:
-        names, grades, prices, localisations, links = getHotels()
-        appendToCsv(names, grades, prices, links, localisations, "bookingCom.csv")
+        cf.appendToCsv(getHotels(), "bookingCom.csv")
 
         changePage()
